+++ client/index.js
@@ -47,58 +47,29 @@ Component.Gravatar = {
     alert('Say hello!');
   },
 
-  // Optional method to initialize the component.
-  // If your component is a constructor/class then
-  // the constructor will be called first before this method.
-  init: function () {
-    // This will be called to initialize the component.
-    // The parent property will be set to the parent
-    // component in the component hierarchy. Whereas
-    // the children, refs and templateInstance properties
-    // will only be set to their defaults (empty array,
-    // empty object and undefined respectively).
-    //
-    // Mixins are guarnteed to have been already initialized so
-    // you safely use callFirstWith() or getFirstWith().
+  initialize: function () {
+    console.log('Gravatar#initialize', this.parent);
   },
 
-  // Optional
-  onCreated: function () {
-    // onCreated called when template has been created.
-    // This is called as a method on your component instance
-    // if it exists. At this stage this.templateInstance will be
-    // defined on the component, but also the entire template instance
-    // API is defined on the component for your convenience.
-    // Each property on the template instance including fields
-    // are methods on the component of the same name.
-    //
-    // Mixins are guarnteed to have had their onCreated methods called first.
+  ready: function () {
+    console.log('Gravatar#ready', this.refs.avatar);
   },
 
-  // Optional
-  onRendered: function () {
-    // onRendered called when template has been inserted
-    // into the DOM. When the HTML for the template has
-    // been materialized and inserted into the DOM any
-    // elements that have a 'ref' or 'data-ref' attribute
-    // will be referenced in 'this.refs'.
-    console.log('Gravatar#onRendered', this.refs.avatar);
-    // By this point the 'children' property will have all
-    // child components inserted (if your component is a parent).
-    //
-    // Mixins are guaranteed to have had their onRendered methods called first.
-  },
-
-  // Optional
-  onDestroyed: function () {
-    // onDestroyed called when template has been removed
-    // from the DOM without being re-rendered. All built-in
-    // properties (parent, children, refs and templateInstance)
-    // are still valid at this point, but will be set to null
-    // after your onDestroyed() method has been called.
-    //
-    // Mixins are guaranteed to have had their onDestroyed methods called first.
+  destroy: function () {
+    console.log('Gravatar#destroy');
   }
 };
 
-Component.Shell = {};
+Component.Shell = class {
+  initialize() {
+    console.log('Shell#initialize');
+  }
+
+  ready() {
+    console.log('Shell#ready');
+  }
+
+  destroy() {
+    console.log('Shell#destroy');
+  }
+}